misc: Switching to audio sprite and m4a + ogg for mobile browsers.
Using an 'audio sprite' and a bit of JS to minimize the data that needs to be loaded + make switching audio tracks
faster feeling. Since MP3 support is hit and miss across mobile browsers, I thought it made sense to use an
m4a(everything except FF) and an ogg(for Firefox).